Refactor KV-cache export/runtime cleanup and document AOTI flows
* extract and simplify HSTU KV-cache export wrapper
* remove stale debug instrumentation and export bring-up prints
* standardize macro-gated logging in KV-cache runtime/binding/client code
* rename AOTI ordering placeholder tensors consistently
* clean C++ replay demo logging and CUDA check behavior
* document end-to-end HSTU KV-cache AOTI C++ and Triton Server workflows
